pyproject.toml
[tool.poetry]
name = "telegram-bot-for-fastapi-backend"
version = "0.1.0"
description = ""
authors = ["Tonmecool <qpwoei.86@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
python-telegram-bot = "^21.2"
django-environ = "^0.11.2"
pydantic = "^2.7.3"
pydantic-settings = "^2.3.0"
dishka = "^1.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
